+++ apps/frontend-pwa/src/pages/bookmarks.tsx
@@ -1,7 +1,7 @@
 import { useQuery } from '@apollo/client'
 import { faRepeat } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { GetUserCoursesDocument } from '@klicker-uzh/graphql/dist/ops'
+import { GetParticipantCoursesDocument } from '@klicker-uzh/graphql/dist/ops'
 import { Button, H1 } from '@uzh-bf/design-system'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
@@ -10,20 +10,18 @@ import Layout from '../components/Layout'
 
 function Bookmarks() {
   const router = useRouter()
-  const { data } = useQuery(GetUserCoursesDocument)
-  console.log(data?.userCourses?.map((course) => course.displayName))
+  const { data } = useQuery(GetParticipantCoursesDocument)
 
   return (
     <Layout courseName="KlickerUZH" displayName="Kursübersicht">
       <div className="flex flex-col gap-2 md:w-full md:max-w-xl md:p-8 md:mx-auto md:border md:rounded">
         <H1 className={{ root: 'text-xl' }}>Kurse</H1>
-        {data?.userCourses?.map((course) => (
+        {data?.participantCourses?.map((course) => (
           <Link
             key={course.id}
-            href={`/bookmarks/${course.id}/course/${course.id}`}
+            href={`/course/${course.id}/bookmarks/${course.id}`}
             legacyBehavior
           >
-            <div>{course.displayName}</div>
             <Button
               className={{
                 root: twMerge(
